Move deck select screen to own file
